docker/requirements.txt
fastapi
wheel
uvicorn
cryptography
pycryptodome
websockets
requests
pyjwt
petlib
passlib[argon2]
pydantic
